requirements.txt
Click==7.1.2
pyusb==1.1.1
Pillow==8.2.0
